split implementation into logical bits
