+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/kibana.jsonc
