+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/src/components/data_cascade_impl/data_cascade_row/index.ts
@@ -7,4 +7,12 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { getESQLStatsGroupByColumnsFromQuery } from './parse_esql';
+export { CascadeRow } from './cascade_row';
+export type { CascadeRowProps } from './cascade_row';
+export { CascadeRowCell } from './cascade_row_cell';
+export type { CascadeRowCellProps } from './cascade_row_cell';
+export {
+  getCascadeRowLeafDataCacheKey,
+  getCascadeRowNodePath,
+  getCascadeRowNodePathValueRecord,
+} from './utils';
