+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/src/components/data_cascade_impl/group_selection_combobox/index.ts
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { getESQLStatsGroupByColumnsFromQuery } from './parse_esql';
+export { SelectionDropdown } from './selection_dropdown';
