+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/src/components/data_cascade_impl/index.ts
@@ -7,4 +7,5 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { DataCascade } from './data_cascade';
+export { DataCascadeImpl } from './data_cascade_impl';
+export type { DataCascadeImplProps } from './data_cascade_impl';
