+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/src/components/index.tsx
@@ -7,4 +7,17 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { DataCascade } from './data_cascade';
+import React, { type ComponentProps } from 'react';
+import { DataCascadeImpl, DataCascadeImplProps } from './data_cascade_impl';
+import { DataCascadeProvider, type GroupNode, type LeafNode } from './data_cascade_provider';
+
+export function DataCascade<N extends GroupNode = GroupNode, L extends LeafNode = LeafNode>({
+  cascadeGroups,
+  ...props
+}: DataCascadeImplProps<N, L> & ComponentProps<typeof DataCascadeProvider>) {
+  return (
+    <DataCascadeProvider cascadeGroups={cascadeGroups}>
+      <DataCascadeImpl<N, L> {...props} />
+    </DataCascadeProvider>
+  );
+}
