+++ src/platform/packages/shared/shared-ux/document_data_cascader/impl/src/index.ts
@@ -7,5 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { DataCascadeProvider } from './lib/store';
 export { DataCascade } from './components';
